Model generator update
