+++ models/__init__.py
@@ -1,6 +1,4 @@
 from .soda import SODa
-from .backbone.vgg import VGGBackbone
-from .backbone.resnet import ResNetBackbone
-from .neck.ssd import SSDNeck
-
-__all__ = "SODa", "VGGBackbone", "SSDNeck", "ResNetBackbone"
+from .backbone import BackboneGen
+from .neck import NeckGen
+__all__ = "SODa", "BackboneGen", "NeckGen"
